added gps
